requirements.txt
kivymd
kivy
datetime
asyncio
telethon
pycryptodome


